requirements.txt
httpx
pyyaml
websockets
backoff
git+https://github.com/Rapptz/discord.py.git@master
jishaku